Switched size() for ssize(), and revamped rendering to use SDL float precison rendering

+++ src/timer.cpp
@@ -19,6 +19,6 @@ float Timer::getSeconds()
     Uint64 currentTimePoint = SDL_GetPerformanceCounter();
 
     // Find the seconds between the current time and past time point
-    float deltaTime = static_cast<float>(currentTimePoint - prevTimePoint) / static_cast<float>(SDL_GetPerformanceFrequency());
+    float deltaTime = static_cast<float>(static_cast<double>(currentTimePoint - prevTimePoint) / static_cast<double>(SDL_GetPerformanceFrequency()));
     return deltaTime;
 }
